Move contact page to the app directory
