+++ src/app/contact/DISCORD_LINK.ts
@@ -1,0 +1,1 @@
+export const DISCORD_LINK = "https://discord.gg/gVxGUERQTq";
